+++ NEWS.md
@@ -1,5 +1,7 @@
 # stringr (development version)
 
+* Better error message if you supply a non-string pattern (#378).
+
 * `str_pad()` gains `use_length` argument to control whether to use the total code
   point width or the number of code points as "width" of a string (#190).